requirements.txt
pandas>=2.0.0
numpy>=1.24.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
groq>=0.36.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
